samples: dtm: nrf5340_cpunet do not redefine uart0 as ipc_uart
Making the `uart0` a `compatible = "nordic,nrf-ipc-uart";` does
not release the `uart0` for other purposes. On nrf5340_cpunet when
when an overlay that uses TWIM0 is used the fact that `uart0` node
is still used (regardless that redefined to use ipc uart not the
hardware uart driver) leads to compilation error.

To overcome this issue the `uart0` is no more redefined but a new
device tree node `uart_ipc` is provided. From application's point
of view this is an additional uart that is used by dtm. The
`uart0` present in device tree can be used for other purpose or
disabled.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>