dfu: dfu_target: Modem delta needs to deinitialize before reset
If the dfu_target_write is called at least once for the modem
delta, it sets the modem into some kind of download mode which
must be stopped before we call dfu_target_reset().

This is not how dfu_target.h is documenting the process, so handle
this extra call to dfu_target_done() internally.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>